+++ src/main/java/spring/certification/aop/package-info.java
@@ -9,8 +9,8 @@
  * 1.5. What two problems arise if you don't solve a cross cutting concern via AOP? {@link spring.certification.aop.q001.AOPConcepts link}<br>
  * 2. What is a pointcut, a join point, an advice, an aspect, a weaving? {@link spring.certification.aop.q002.AOPTerminology link}<br>
  * 3. How does Spring framework solve a cross cutting concern? {@link spring.certification.aop.q003.CrossCuttingSolution link}<br>
- * 4.1. What are the limitations of the two proxy types? {@link spring.certification.aop.Q004limitations link}<br>
- * 4.2. What visibility must Spring bean methods have to be proxied using Spring AOP? {@link spring.certification.aop.Q004limitations link}<br>
+ * 4.1. What are the limitations of the two proxy types? {@link spring.certification.aop.q004.ProxyLimitations link}<br>
+ * 4.2. What visibility must Spring bean methods have to be proxied using Spring AOP? {@link spring.certification.aop.q004.ProxyLimitations link}<br>
  * 5.1. How many advice types does Spring support. Can you name each one? {@link spring.certification.aop.Q005advice link}<br>
  * 5.2. What different advice types are used for? {@link spring.certification.aop.Q005advice link}<br>
  * 5.3. Which types of advices can you use if you would like to try and catch exceptions? {@link spring.certification.aop.Q005advice link}<br>
