+++ src/main/java/spring/certification/ioc/q024/SpringAOP.java
@@ -2,7 +2,7 @@ package spring.certification.ioc.q024;
 
 import spring.certification.aop.q001.AOPConcepts;
 import spring.certification.aop.q003.CrossCuttingSolution;
-import spring.certification.aop.Q004limitations;
+import spring.certification.aop.q004.ProxyLimitations;
 
 /**
  * Spring v5.0 Professional Certification.
@@ -15,7 +15,7 @@ import spring.certification.aop.Q004limitations;
  * Answer:<br>
  * AOP implementation details as well as different types of proxy objects are discussed in the {@link CrossCuttingSolution proxy}
  * question of the AOP certification section.<br>
- * Limitations are discussed in the {@link Q004limitations limitations} question of the AOP certification section.<br>
+ * Limitations are discussed in the {@link ProxyLimitations limitations} question of the AOP certification section.<br>
  * Advantages of using proxies are discussed in the {@link AOPConcepts consepts} question of the AOP certification section.<br>
  *
  * @author Valentine Shemyako
