+++ src/main/java/spring/certification/security/Q014annotationworkings.java
@@ -1,7 +1,7 @@
 package spring.certification.security;
 
 import spring.certification.aop.q003.CrossCuttingSolution;
-import spring.certification.aop.Q004limitations;
+import spring.certification.aop.q004.ProxyLimitations;
 
 /**
  * Spring v5.0 Professional Certification.
@@ -12,7 +12,7 @@ import spring.certification.aop.Q004limitations;
  * Answer:<br>
  * Method security annotations are implemented using AOP proxies. AOP proxy is a dynamically created
  * object which wraps target object and intercepts method invocation. Please refer to the AOP section
- * for explanation of this {@link CrossCuttingSolution approach} and its {@link Q004limitations limitations}.<br>
+ * for explanation of this {@link CrossCuttingSolution approach} and its {@link ProxyLimitations limitations}.<br>
  *
  * @author Valentine Shemyako
  * @since January 18, 2019
